Created simple role and user management through identity provider.
Added User and Role Services.
Created initial Migration scripts.

+++ Timesheets.Domain/Models/Project.cs
@@ -12,7 +12,7 @@ namespace Timesheets.Domain.Models
         public IList<ProjectDepartment> RelatedDeparments { get; set; }
 
         // one project is owned by one deparment
-        public long OwnerDeparmentId { get; set; }
+        public int OwnerDeparmentId { get; set; }
         public Department OwnerDeparment { get; set; }
     }
 }
